backend/requirements.txt
requests==2.31.0
python-dotenv==1.0.0
apscheduler==3.10.4
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
pg8000==1.30.3


